// Codigo/front-end/src/app/interfaces/Teacher.ts
export interface Teacher {
  id: number;
  fullName: string;
  cpf: string;
  rg: string;
  email: string;
  phoneNumber: string;
  addressNumber: string;
  addressComplement: string;
  streetAddress: string;
  neighborhood: string;
  zipCode: string;
  cityOfResidence: string;
  homeState: string;
  registrationDate: string;
  birthDate: string;
  salary: number;
  hireDate: string;
  terminationDate: string;
  academicFormation: string;
  academicFormationStatus: string;
}